+++ src/integrations/mod.rs
@@ -1,2 +1,5 @@
+#[cfg(feature = "bytemuck")]
+mod bytemuck;
+
 #[cfg(feature = "serde")]
 mod serde;